feat: support loading custom config via settings
Add a `commitlint.config.file` configuration to allow the path to the
commitlint config to be specified via settings, rather than
automatically detected.

Enhance error handling in the event the specified path can't be loaded,
by updating the status icon with a new "error" state and logging a
somewhat more specific message to the extension's output log.

Fixes #20