fix: use workspace config for non-file schemes
For documents whose uri scheme is something other than `file`, fall back
to a configuration based on the workspace rather than trying to locate
one based on the file's path. Paths corresponding to non-`file` schemes
would not expected to be traversable, so no configuration is likely to
be found anyway.

This behavior is equivalent to the existing behavior for any untitled
document (e.g., opening a new file and manually setting its mode to "git
commit message").

Fixes #69