fix: prefer local library path as PREFIX
When loading a local instance of `@commitlint/load`, use its path as the
`PREFIX` when attempting to actually load a configuration. This should
improve the chances that any installed configurations or plugins are
discovered by preferring the location of a previously discovered library
rather over a more generic global location.

Fixes #600
Fixes #744