fix: support verbose commits
Verbose commits append additional data to a commit's default message,
but also use a special "scissors line" to delineate the end of the
author's commit message. Content following this line is not part of the
submitted message, and so should not be linted.

Verbose commits may be enabled via a local or global git config, or be
enabled (or disabled) via command line flags passed to `git commit`. The
SCM input implicitly (if unexpectedly) respects the git config. However,
there is not a reliable way to determine whether the scissors line
should be honored when the full editor is open, particularly if a
command line flag was used to override the prevailing config.

As a heuristic for full editors (i.e., not the SCM input), track for
each document whether it was initially observed to contain a scissors
line. If it did, it is probable that it was populated by git with
verbose mode enabled (via any means), so the scissors line should be
honored. If not, verbose mode is probably not active, so the scissors
line should not receive special treatment.

Fixes #593